+++ apps/meteor/server/methods/removeRoomModerator.ts
@@ -9,6 +9,7 @@ import { Meteor } from 'meteor/meteor';
 import { hasPermissionAsync } from '../../app/authorization/server/functions/hasPermission';
 import { notifyOnSubscriptionChangedById } from '../../app/lib/server/lib/notifyListener';
 import { settings } from '../../app/settings/server';
+import { syncRoomRolePriorityForUserAndRoom } from '../lib/roles/syncRoomRolePriority';
 
 declare module '@rocket.chat/ddp-client' {
 	// eslint-disable-next-line @typescript-eslint/naming-convention
@@ -66,6 +67,8 @@ Meteor.methods<ServerMethods>({
 		}
 
 		const removeRoleResponse = await Subscriptions.removeRoleById(subscription._id, 'moderator');
+		await syncRoomRolePriorityForUserAndRoom(userId, rid, subscription.roles?.filter((r) => r !== 'moderator') || []);
+
 		if (removeRoleResponse.modifiedCount) {
 			void notifyOnSubscriptionChangedById(subscription._id);
 		}