+++ packages/model-typings/src/models/IRoomsModel.ts
@@ -313,4 +313,6 @@ export interface IRoomsModel extends IBaseModel<IRoom> {
 	countByCreatedOTR(options?: CountDocumentsOptions): Promise<number>;
 	countByBroadcast(options?: CountDocumentsOptions): Promise<number>;
 	countByE2E(options?: CountDocumentsOptions): Promise<number>;
+	markRolePrioritesCreatedForRoom(rid: IRoom['_id']): Promise<UpdateResult>;
+	hasCreatedRolePrioritiesForRoom(rid: IRoom['_id']): Promise<number>;
 }