+++ packages/core-services/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rocket.chat/core-services
 
+## 0.7.4
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.2.1
+  - @rocket.chat/rest-typings@7.2.1
+  - @rocket.chat/models@1.1.1
+  </details>
+
 ## 0.7.3
 
 ### Patch Changes