+++ packages/gazzodown/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @rocket.chat/gazzodown
 
+## 14.0.1
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.2.1
+  - @rocket.chat/ui-contexts@14.0.1
+  - @rocket.chat/ui-client@14.0.1
+  </details>
+
 ## 14.0.0
 
 ### Patch Changes